+++ .gitignore
@@ -1,2 +1,3 @@
 /node_modules
-/db/db.json
+**/db.json
+/dist/*
