+++ package.json
@@ -2,9 +2,11 @@
   "name": "vacancy-graphql",
   "version": "1.0.0",
   "description": "Vacancy GraphQL Service",
-  "main": "index.js",
+  "main": "dist/index",
+  "types": "dist/index",
   "scripts": {
-    "start": "nodemon --ignore db.json index.js",
+    "start": "nodemon --ignore ./src/db/db.json --exec ts-node ./src/index.ts",
+    "build": "tsc",
     "lint": "eslint **/*.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -14,13 +16,22 @@
   "author": "Eugene Lemesh",
   "license": "ISC",
   "devDependencies": {
+    "@types/cookie-parser": "^1.4.1",
+    "@types/express": "^4.16.0",
+    "@types/express-session": "^1.15.10",
+    "@types/graphql": "^0.13.3",
+    "@types/node": "^10.5.1",
+    "@types/passport": "^0.4.5",
+    "@types/passport-local": "^1.0.33",
     "eslint": "^5.0.1",
     "eslint-config-standard": "^11.0.0",
     "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-node": "^6.0.1",
     "eslint-plugin-promise": "^3.8.0",
     "eslint-plugin-standard": "^3.1.0",
-    "nodemon": "^1.17.5"
+    "nodemon": "^1.17.5",
+    "ts-node": "^7.0.0",
+    "typescript": "^2.9.2"
   },
   "dependencies": {
     "apollo-server-express": "^1.3.6",
