+++ src/auth.ts
@@ -1,11 +1,13 @@
-const passport = require('passport')
-const LocalStrategy = require('passport-local').Strategy
+import passport from 'passport'
+import { Strategy } from 'passport-local'
+import { Application } from 'express'
+
 const user = {
   username: 'user',
   password: 'qwerty123'
 }
 
-passport.use(new LocalStrategy({}, (username, password, done) => {
+passport.use(new Strategy({}, (username, password, done) => {
   if (user.username === username.toLowerCase() && user.password === password) {
     done(null, user)
   } else {
@@ -13,7 +15,7 @@ passport.use(new LocalStrategy({}, (username, password, done) => {
   }
 }))
 
-passport.serializeUser((user, done) => {
+passport.serializeUser((user: any, done) => {
   done(null, user.username)
 })
 
@@ -21,7 +23,7 @@ passport.deserializeUser((username, done) => {
   done(null, user)
 })
 
-module.exports = app => {
+export default (app: Application) => {
   app.use(passport.initialize())
   app.use(passport.session())
 }
